Fix imageSize warning

+++ Source/GUI/Components/PoZePlot.cpp
@@ -292,7 +292,7 @@ juce::MouseCursor PoZePlot::Point::getRotatedMagnitudeCursor (const float angle)
     // Max size is at 45 degrees, with size = side * sqrt(2)
     const float imageSize = std::round (std::sqrt (2.0f) * maxDim);
 
-    juce::Image image (juce::Image::ARGB, imageSize, imageSize, true);
+    juce::Image image (juce::Image::ARGB, (int)imageSize, (int)imageSize, true);
 
     {
         juce::Graphics g (image);
@@ -310,7 +310,7 @@ juce::MouseCursor PoZePlot::Point::getRotatedMagnitudeCursor (const float angle)
         svg->draw (g, 1.0f, transform);
     }
 
-    return juce::MouseCursor { image,0,0,6.0f };
+    return juce::MouseCursor { image, 0, 0, 6.0f };
 }
